Add files via upload
There are three files required in order to run yolo in openCV, coco names file, yolo config file and the weights file. The yolo weights files is large and hence can't be uploaded. Just google it and you will find it. Here, in this project we are using yolov3. The citation of the paper on YOLOv3,
 @article{yolov3,
  title={YOLOv3: An Incremental Improvement},
  author={Redmon, Joseph and Farhadi, Ali},
  journal = {arXiv},
  year={2018}
}